+++ oxc/oxlint.js
@@ -1,0 +1,7 @@
+import { defineConfig } from 'oxlint';
+import base from './base.js';
+import react from './react.js';
+
+export default defineConfig({
+  extends: [base, react],
+});