+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased]
+#### Added
+* Add configurations for Oxlint.
+
+#### Changed
+* *Nothing*
+
+#### Deprecated
+* *Nothing*
+
+#### Removed
+* *Nothing*
+
+#### Fixed
+* *Nothing*
+
+
 ## [3.8.0] - 2026-02-08
 #### Added
 * Add support for `eslint` 10.